# data/contracts/openapi.v2.json
{
  "info": {
    "title": "Toonflow Provider API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/jobs/{id}": {
      "get": {
        "operationId": "getGenerationJob",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    },
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "cancelReason": {
                          "type": "string"
                        },
                        "cancelRequestedAt": {
                          "maximum": 9007199254740991,
                          "minimum": -9007199254740991,
                          "type": "integer"
                        },
                        "canonicalModelId": {
                          "pattern": "^[a-z0-9]+(?:[._:-][a-z0-9]+)*$",
                          "type": "string"
                        },
                        "consumer": {
                          "oneOf": [
                            {
                              "additionalProperties": false,
                              "properties": {
                                "context": {
                                  "additionalProperties": false,
                                  "properties": {
                                    "projectId": {
                                      "exclusiveMinimum": 0,
                                      "maximum": 9007199254740991,
                                      "type": "integer"
                                    },
                                    "scriptId": {
                                      "maximum": 9007199254740991,
                                      "minimum": 0,
                                      "type": "integer"
                                    },
                                    "trackId": {
                                      "exclusiveMinimum": 0,
                                      "maximum": 9007199254740991,
                                      "type": "integer"
                                    }
                                  },
                                  "required": [
                                    "projectId",
                                    "scriptId",
                                    "trackId"
                                  ],
                                  "type": "object"
                                },
                                "key": {
                                  "maxLength": 200,
                                  "minLength": 1,
                                  "type": "string"
                                },
                                "type": {
                                  "const": "workbench",
                                  "type": "string"
                                }
                              },
                              "required": [
                                "type",
                                "key",
                                "context"
                              ],
                              "type": "object"
                            },
                            {
                              "additionalProperties": false,
                              "properties": {
                                "context": {
                                  "additionalProperties": false,
                                  "properties": {
                                    "assetId": {
                                      "exclusiveMinimum": 0,
                                      "maximum": 9007199254740991,
                                      "type": "integer"
                                    },
                                    "assetType": {
                                      "enum": [
                                        "role",
                                        "scene",
                                        "tool"
                                      ],
                                      "type": "string"
                                    },
                                    "projectId": {
                                      "exclusiveMinimum": 0,
                                      "maximum": 9007199254740991,
                                      "type": "integer"
                                    }
                                  },
                                  "required": [
                                    "projectId",
                                    "assetId",
                                    "assetType"
                                  ],
                                  "type": "object"
                                },
                                "key": {
                                  "maxLength": 200,
                                  "minLength": 1,
                                  "type": "string"
                                },
                                "type": {
                                  "const": "asset_image",
                                  "type": "string"
                                }
                              },
                              "required": [
                                "type",
                                "key",
                                "context"
                              ],
                              "type": "object"
                            }
                          ]
                        },
                        "continuation": {
                          "additionalProperties": false,
                          "properties": {
                            "parentJobId": {
                              "format": "uuid",
                              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                              "type": "string"
                            }
                          },
                          "required": [
                            "parentJobId"
                          ],
                          "type": "object"
                        },
                        "createdAt": {
                          "maximum": 9007199254740991,
                          "minimum": -9007199254740991,
                          "type": "integer"
                        },
                        "error": {},
                        "id": {
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                          "type": "string"
                        },
                        "idempotencyKey": {
                          "type": "string"
                        },
                        "input": {
                          "additionalProperties": false,
                          "properties": {
                            "assets": {
                              "items": {
                                "additionalProperties": false,
                                "properties": {
                                  "assetId": {
                                    "minLength": 1,
                                    "type": "string"
                                  },
                                  "durationSeconds": {
                                    "minimum": 0,
                                    "type": "number"
                                  },
                                  "kind": {
                                    "enum": [
                                      "image",
                                      "video",
                                      "audio"
                                    ],
                                    "type": "string"
                                  },
                                  "role": {
                                    "minLength": 1,
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "assetId",
                                  "kind",
                                  "role"
                                ],
                                "type": "object"
                              },
                              "type": "array"
                            },
                            "mode": {
                              "minLength": 1,
                              "type": "string"
                            },
                            "values": {
                              "additionalProperties": {},
                              "propertyNames": {
                                "type": "string"
                              },
                              "type": "object"
                            }
                          },
                          "required": [
                            "values",
                            "assets"
                          ],
                          "type": "object"
                        },
                        "nextRunAt": {
                          "maximum": 9007199254740991,
                          "minimum": -9007199254740991,
                          "type": "integer"
                        },
                        "offeringId": {
                          "pattern": "^[a-z0-9]+(?:[._:-][a-z0-9]+)*$",
                          "type": "string"
                        },
                        "operation": {
                          "enum": [
                            "language.generate",
                            "language.stream",
                            "image.generate",
                            "image.edit",
                            "video.generate",
                            "video.status",
                            "video.cancel",
                            "files.upload",
                            "search.ground"
                          ],
                          "type": "string"
                        },
                        "pollAttemptCount": {
                          "maximum": 9007199254740991,
                          "minimum": 0,
                          "type": "integer"
                        },
                        "providerId": {
                          "pattern": "^[a-z0-9]+(?:[._:-][a-z0-9]+)*$",
                          "type": "string"
                        },
                        "providerOutcome": {
                          "enum": [
                            "unknown",
                            "queued",
                            "running",
                            "succeeded",
                            "failed",
                            "cancelled"
                          ],
                          "type": "string"
                        },
                        "requiresReconciliation": {
                          "type": "boolean"
                        },
                        "result": {},
                        "schemaVersion": {
                          "type": "string"
                        },
                        "state": {
                          "enum": [
                            "queued",
                            "preparing_assets",
                            "submitting",
                            "submitted",
                            "remote_queued",
                            "running",
                            "importing",
                            "submission_unknown",
                            "succeeded",
                            "failed",
                            "cancelled",
                            "abandoned"
                          ],
                          "type": "string"
                        },
                        "updatedAt": {
                          "maximum": 9007199254740991,
                          "minimum": -9007199254740991,
                          "type": "integer"
                        },
                        "version": {
                          "maximum": 9007199254740991,
                          "minimum": 0,
                          "type": "integer"
                        }
                      },
                      "required": [
                        "id",
                        "schemaVersion",
                        "idempotencyKey",
                        "canonicalModelId",
                        "offeringId",
                        "providerId",
                        "operation",
                        "input",
                        "state",
                        "nextRunAt",
                        "pollAttemptCount",
                        "version",
                        "createdAt",
                        "updatedAt",
                        "requiresReconciliation"
                      ],
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Generation job snapshot"
          }
        }
      }
    },
...
  }
}